update niswitch helpers
